requirements.txt
Flask
gunicorn
pdf2docx
pdf2image
pytesseract
python-docx
PyMuPDF
opencv-python-headless
